Use GitHub Pull Request reference for merging
Rather than fetching the forks, this fetches the remote associated with the
pull request as described
https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/checking-out-pull-requests-locally.
In the case of deleted forks, this should improve the robustness of the merge
command. It might also result in slight performance improvements as only the
relevant commits will be fetches rather than all the branches of the forks.