Refactor status filter logic to handle both Status and Checks API
As the CI/CD of our organizational repositories are getting migrated to using
GitHub Actions, changes are required at the level of scc to handle the output
of these runs. Unlike Travis CI which uses the Status API, GitHub Actions use
the recent Check API in order to report the outcome.

This PR makes the minimal amount of change to re-use the existing semantics
of the merge command but apply the status filter to both existing check_suites
as well as the status.

The status reading logic is also updated to use the count of `get_statuses()`
to determine whether a status exists and then use the combined status API to
return a single state.

The check-suite logic uses the count of `get_check_suites()` to determine
whether a check-suite has been executed and the tests the conclusion of each
check-suite.

The state filtering is also updated to handle the `skipped` conclusion as a
successful one.