Filter check-suites by app ID to only verify the conclusion of GitHub actions
Other apps (GitHub pages, dependabot) create their own check suites which might
not set the conclusion field used for testing the inclusion of PRs. In a first
instance, this commit restricts the status check to the GitHub actions check
suites only. If this needs to be expanded, we could look into making this
parametrizable.